openvino/all: added component for C API
