+++ recipes/openvino/all/test_package/conanfile.py
@@ -45,5 +45,8 @@ class TestPackageConan(ConanFile):
 
     def test(self):
         if can_run(self):
-            bin_path = os.path.join(self.cpp.build.bindirs[0], "test_package")
-            self.run(bin_path, env="conanrun")
+            bin_path_c = os.path.join(self.cpp.build.bindirs[0], "test_package_c")
+            self.run(bin_path_c, env="conanrun")
+
+            bin_path_cpp = os.path.join(self.cpp.build.bindirs[0], "test_package_cpp")
+            self.run(bin_path_cpp, env="conanrun")
