+++ recipes/openvino/all/test_package/test_package.cpp
@@ -10,41 +10,6 @@
     if ((statement) == 0) \
         return 1;
 
-int test_available_devices() {
-    ov_core_t* core = NULL;
-    char* ret = NULL;
-    OV_SUCCESS(ov_core_create(&core));
-#ifdef ENABLE_INTEL_CPU
-    OV_SUCCESS(ov_core_get_property(core, "CPU", "AVAILABLE_DEVICES", &ret));
-#else
-    OV_FAIL(ov_core_get_property(core, "CPU", "AVAILABLE_DEVICES", &ret));
-#endif
-#ifdef ENABLE_INTEL_GPU
-    OV_SUCCESS(ov_core_get_property(core, "GPU", "AVAILABLE_DEVICES", &ret));
-#else
-    OV_FAIL(ov_core_get_property(core, "GPU", "AVAILABLE_DEVICES", &ret));
-#endif
-#ifdef ENABLE_AUTO
-    OV_SUCCESS(ov_core_get_property(core, "AUTO", "SUPPORTED_METRICS", &ret));
-    OV_SUCCESS(ov_core_get_property(core, "MULTI", "SUPPORTED_METRICS", &ret));
-#else
-    OV_FAIL(ov_core_get_property(core, "AUTO", "SUPPORTED_METRICS", &ret));
-    OV_FAIL(ov_core_get_property(core, "MULTI", "SUPPORTED_METRICS", &ret));
-#endif
-#ifdef ENABLE_HETERO
-    OV_SUCCESS(ov_core_get_property(core, "HETERO", "SUPPORTED_METRICS", &ret));
-#else
-    OV_FAIL(ov_core_get_property(core, "HETERO", "SUPPORTED_METRICS", &ret));
-#endif
-#ifdef ENABLE_AUTO_BATCH
-    OV_SUCCESS(ov_core_get_property(core, "BATCH", "SUPPORTED_METRICS", &ret));
-#else
-    OV_FAIL(ov_core_get_property(core, "BATCH", "SUPPORTED_METRICS", &ret));
-#endif
-    ov_core_free(core);
-    return 0;
-}
-
 int test_available_frontends() {
     ov::frontend::FrontEndManager manager;
     auto frontend_found = [&] (const std::string & name) -> int {
@@ -90,7 +55,6 @@ int test_available_frontends() {
 }
 
 int main() {
-    OV_SUCCESS(test_available_devices());
     OV_SUCCESS(test_available_frontends());
 
     // Deinitialize OpenVINO. Important for old systems like Ubuntu 16.04 with obsolete glibc,
